requirements.txt
# Core Web Framework
flask>=2.0.0
flask-cors>=4.0.0
werkzeug>=2.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.20.0,<2.0

# LLM and AI Components
langchain>=0.1.0
langchain-ibm>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Vector Database (Elasticsearch)
elasticsearch>=8.0.0
sentence-transformers>=2.2.0

# Utilities
urllib3>=1.26.0

# Optional: Production Server
# gunicorn>=20.1.0